+++ src/cli/mod.rs
@@ -24,7 +24,7 @@ use thiserror::Error;
 pub use self::build::*;
 pub use self::doc::*;
 pub use self::fmt_project::FmtProjectCommand;
-pub use self::init::*;
+pub use self::init::{InitCommand, InitKind};
 pub use self::plugin::*;
 pub use self::serve::*;
 pub use self::upload::*;
@@ -102,7 +102,6 @@ pub struct ColorChoiceParseError {
 
 #[derive(Debug, StructOpt)]
 pub enum Subcommand {
-    /// Creates a new Rojo project.
     Init(InitCommand),
 
     /// Serves the project's files for use with the Rojo Studio plugin.
@@ -121,66 +120,6 @@ pub enum Subcommand {
 
     /// Manages Rojo's Roblox Studio plugin.
     Plugin(PluginCommand),
-}
-
-/// Initializes a new Rojo project.
-#[derive(Debug, StructOpt)]
-pub struct InitCommand {
-    /// Path to the place to create the project. Defaults to the current directory.
-    #[structopt(default_value = "")]
-    pub path: PathBuf,
-
-    /// The kind of project to create, 'place' or 'model'. Defaults to place.
-    #[structopt(long, default_value = "place")]
-    pub kind: InitKind,
-}
-
-impl InitCommand {
-    pub fn absolute_path(&self) -> Cow<'_, Path> {
-        resolve_path(&self.path)
-    }
-}
-
-/// The templates we support for initializing a Rojo project.
-#[derive(Debug, Clone, Copy)]
-pub enum InitKind {
-    /// A place that matches what File -> New does in Roblox Studio.
-    Place,
-
-    /// An empty model, suitable for a library or plugin.
-    Model,
-}
-
-impl FromStr for InitKind {
-    type Err = InitKindParseError;
-
-    fn from_str(source: &str) -> Result<Self, Self::Err> {
-        match source {
-            "place" => Ok(InitKind::Place),
-            "model" => Ok(InitKind::Model),
-            _ => Err(InitKindParseError {
-                attempted: source.to_owned(),
-            }),
-        }
-    }
-}
-
-/// Error type for failing to parse an `InitKind`.
-#[derive(Debug)]
-pub struct InitKindParseError {
-    attempted: String,
-}
-
-impl Error for InitKindParseError {}
-
-impl fmt::Display for InitKindParseError {
-    fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            formatter,
-            "Invalid init kind '{}'. Valid kinds are: place, model",
-            self.attempted
-        )
-    }
 }
 
 /// Expose a Rojo project through a web server that can communicate with the