+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Added the `gameId` and `placeId` optional properties to project files.
     * When connecting from the Rojo Roblox Studio plugin, Rojo will set the game and place ID of the current place to these values, if set.
     * This is equivalent to running `game:SetUniverseId(...)` and `game:SetPlaceId(...)` from the command bar in Studio. 
+* Fixed `Name` and `Parent` properties being allowed in Rojo projects. ([#413](pr-413))
+
+[pr-413]: https://github.com/rojo-rbx/rojo/pull/413
 
 ## [7.0.0-alpha.3][7.0.0-alpha.3] (February 19, 2021)
 * Updated dependencies, fixing `OptionalCoordinateFrame`-related issues.