+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * When connecting from the Rojo Roblox Studio plugin, Rojo will set the game and place ID of the current place to these values, if set.
     * This is equivalent to running `game:SetUniverseId(...)` and `game:SetPlaceId(...)` from the command bar in Studio. 
 * Fixed `Name` and `Parent` properties being allowed in Rojo projects. ([#413](pr-413))
+* Added "EXPERIMENTAL!" label to two-way sync toggle in Rojo's Roblox Studio plugin.
 
 [pr-413]: https://github.com/rojo-rbx/rojo/pull/413