+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Rojo Changelog
 
 ## Unreleased Changes
+* Added "EXPERIMENTAL!" label to two-way sync toggle in Rojo's Roblox Studio plugin.
 
 ## [6.1.0][6.1.0] (April 12, 2021)
 * Updated dependencies, fixing OptionalCoordinateFrame-related issues.